Tons of stuff - My: 'I hate scripting & would never want to re-write all of these' "release"
mostly nfs helper scripts; playlist conversion; and misc improvements &, of course, *tons of bug fixes*

p.s. I'm not a fan of scripting - actually I pretty much find it a neccassary evil.  LOL exactly why I'm commiting these fixes.  Oh f* I'd hate to loose all of these.

Committed on : 2009-10-08 @ 07:01:49 PM

+++ media/gpodder-helper-scripts/gPodder:Refetch.tcsh
@@ -15,24 +15,43 @@ if( ${?1} && ( "${1}" == "--silent" || "${1}" == "--quiet" ) ) then
 	shift;
 endif
 
+if( ${?1} && ( "${1}" == "--verbose" || "${1}" == "--debug" ) ) then
+	set debug_enabled;
+	shift;
+endif
+
 set mp3_player_folder="`grep 'mp3_player_folder' '${HOME}/.config/gpodder/gpodder.conf' | cut -d= -f2 | cut -d' ' -f2`";
 
 cd "${mp3_player_folder}";
-set search_attribute="`echo "\""${1}"\"" | sed 's/\(\-\-[^=]\+\)=\(.*\)/\1/'`";
-set search_value="`echo "\""${1}"\"" | sed 's/\(\-\-[^=]\+\)=\(.*\)/\2/'`";
 
-foreach podcast_match( "`${search_script} ${1} | cut -d':' -f2 | sed 's/\\!//'`" )
-	set refetch_script="${mp3_player_folder}/gPodder:Refetch:`echo '${podcast_match}' | sed 's/\([\-\=\/\*\?\.\[\]()]\+\)/\:/g'`.tcsh";
+if( ${?GREP_OPTIONS} ) then
+	set grep_options="${GREP_OPTIONS}";
+	unsetenv GREP_OPTIONS;
+endif
+
+if(! ${?eol} ) setenv eol='$';
+
+set search_attribute="`echo "\""${1}"\"" | sed 's/^\-\-\([^=]\+\)=\(.*\)${eol}/\1/'`";
+set search_value="`echo "\""${1}"\"" | sed 's/^\-\-\([^=]\+\)=\(.*\)${eol}/\2/'`";
+
+
+foreach podcast_match( "`${search_script} --${search_attribute}="\""${search_value}"\"" | sed 's/'\''/\\'\''/g' | sed 's/^[^\:]\+:\(.*\)${eol}/\1/g' | sed 's/\\!//'`" )
+	if( ${?debug_enabled} ) echo "${search_value}";
+	set podcast_match="`printf "\""${podcast_match}"\"" | sed 's/\([-\ \(\)]\)/\\\1/g' | sed 's/'\''/\\'\''/g'`";
+	if( ${?debug_enabled} ) echo "${podcast_match}";
+	set refetch_script="${mp3_player_folder}/gPodder:Refetch:`echo "\""${search_value}"\"" | sed 's/\([\-\=\/\*\?\.\[\]()]\+\)/\:/g'`:`echo "\""${podcast_match}"\"" | sed 's/\([\-\=\/\*\?\.\[\]()]\+\)/\:/g'`.tcsh";
+	if( ${?debug_enabled} ) echo "${refetch_script}";
 	
-	${search_script} --verbose ${search_attribute}="${podcast_match}" >! "${refetch_script}.tmp";
-	ex -E '+1,$s/[\r\n]\+//g' '+s/\(<\/item>\)/\1\n/g' '+s/#//g' '+1,$s/.*<title>\([^>]\+\)<\/title>.*<title>\([^<]\+\)<\/title>.*<url>\([^<]\+\)<\/url>.*/if( -d "\1" ) then\relse\r\tmkdir "\1"\rendif\rwget -c -O "\1\/\2" "\3"/g' '+2,$s/^\(wget\ \-c\ \-O\ \)\"\([^\"]\+\)\"\ \"\([^\"]\+\)\.\([^\.\"]\+\)\"$/\1\ \"\2\.\4\"\ \3\.\4/' '+1,$s/\!//g' '+wq' "${refetch_script}.tmp" >& /dev/null;
+	if( ${?debug_enabled} ) echo ${search_script} --verbose --${search_attribute}=\"${podcast_match}\" \>\! \"${refetch_script}.tmp\";
+	${search_script} --verbose --${search_attribute}="${podcast_match}" >! "${refetch_script}.tmp";
+	ex -E -n -X '+1,$s/[\r\n]\+//g' '+s/\(<\/item>\)/\1\n/g' '+s/#//g' '+1,$s/.*<title>\([^>]\+\)<\/title>.*<title>\([^<]\+\)<\/title>.*<url>\([^<]\+\)<\/url>.*/if( -d "\1" ) then\relse\r\tmkdir "\1"\rendif\rwget -c -O "\1\/\2" "\3"/g' '+2,$s/^\(wget\ \-c\ \-O\ \)\"\([^\"]\+\)\"\ \"\([^\"]\+\)\.\([^\.\"]\+\)\"$/\1\ \"\2\.\4\"\ '\''\3\.\4'\''/' '+1,$s/\!//g' '+wq!' "${refetch_script}.tmp" > /dev/null;
 	
 	while ( `/usr/bin/grep --perl-regexp '("[^\/]+)\/(.*)"' "${refetch_script}.tmp"` != "" )
-		ex -E '+1,$s/\("[^\/]\+\)\/\(.*"\)/\1\-\2/g' '+wq' "${refetch_script}.tmp" >& /dev/null;
+		ex -E -n -X '+1,$s/\("[^\/]\+\)\/\(.*"\)/\1\-\2/g' '+wq!' "${refetch_script}.tmp" >& /dev/null;
 	end
 	
-	set podcast_dir=`head -3 "${refetch_script}.tmp" | tail -1 | sed 's/.*mkdir "\([^"]\+\)"/\1/'`;
-	ex -E '+4,$s/\('"${podcast_dir}"'\)\-/\1\//g' '+wq' "${refetch_script}.tmp";
+	set podcast_dir=`head -3 "${refetch_script}.tmp" | tail -1 | sed 's/.*mkdir "\([^"]\+\)"/\1/' | sed 's/\([()]\)/\\\1/g' | sed 's/'\''/\\'\''/g'`;
+	ex -E -n -X '+4,$s/\('"${podcast_dir}"'\)\-/\1\//g' '+wq!' "${refetch_script}.tmp";
 	
 	if( `wc -l "${refetch_script}.tmp" | sed 's/^\([0-9]\+\)\ .*/\1/g'` > 0 ) then
 		printf '#\!/bin/tcsh -f\n' >! "${refetch_script}";
@@ -41,6 +60,11 @@ foreach podcast_match( "`${search_script} ${1} | cut -d':' -f2 | sed 's/\\!//'`"
 		"${refetch_script}" ${silent};
 		if( ${status} == 0 ) rm "${refetch_script}";
 	endif
-	rm "${refetch_script}.tmp";
+	if( ${status} == 0 ) rm "${refetch_script}.tmp";
 end
+
+if( ${?grep_options} ) then
+	setenv GREP_OPTIONS "${grep_options}";
+	unset grep_options;
+endif
